refactor(member): member  모듈 리팩토링 - Member 모듈 리팩토링 - 패키지 구조 정리 및 네이밍 수정
