--- service/member/src/main/java/innercircle/member/application/MemberUseCase.java
@@ -1,8 +1,0 @@
-package innercircle.member.application;
-
-public interface MemberUseCase {
-
-    MemberResponse createMember(MemberCreateRequest request);
-
-
-}
