+++ service/member/src/main/java/innercircle/member/application/port/in/CreateMemberUseCase.java
@@ -1,0 +1,10 @@
+package innercircle.member.application.port.in;
+
+import innercircle.member.application.MemberCreateRequest;
+import innercircle.member.application.MemberResponse;
+
+public interface CreateMemberUseCase {
+
+    MemberResponse createMember(MemberCreateRequest request);
+
+}
